fix: Set `limit` to `uint32_t` instead of `size_t`

+++ src/grpc_client.hpp
@@ -21,7 +21,8 @@ public:
   [[nodiscard]] std::optional<std::string> get(const std::string &key);
   void remove(const std::string &key);
   [[nodiscard]] std::vector<std::pair<std::string, std::string>>
-  scan(const std::string &start_key, const std::string &end_key, size_t limit);
+  scan(const std::string &start_key, const std::string &end_key,
+       uint32_t limit);
 
 private:
   std::unique_ptr<kv::v1::KvStoreService::Stub> stub_;
